requirements.txt
absl-py==0.13.0
boto3==1.17.112
botocore==1.20.112
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.2
click==8.0.1
future==0.18.2
google-auth==1.33.0
google-auth-oauthlib==0.4.4
grpcio==1.38.1
idna==3.2
importlib-metadata==4.6.1
jmespath==0.10.0
joblib==1.0.1
Markdown==3.3.4
numpy==1.21.0
oauthlib==3.1.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynvml==11.0.0
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytorch-transformers==1.2.0
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
sacremoses==0.0.45
scikit-learn==0.23.2
scipy==1.7.0
sentencepiece==0.1.96
six==1.16.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
torch==1.2.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
